requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
curl_cffi==0.12.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-search-results==2.4.2
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.0.1
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.29
requests==2.32.4
requests-file==2.1.0
rpds-py==0.26.0
safetensors==0.5.3
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.65
